tests/wpt/web-platform-tests/css/css-text/parsing/word-boundary-expansion-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-boundary-expansion with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-boundary-expansion">
<meta name="assert" content="word-boundary-expansion supports the full grammar 'none | space | ideographic-space'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("word-boundary-expansion", "none");
test_valid_value("word-boundary-expansion", "space");
test_valid_value("word-boundary-expansion", "ideographic-space");
</script>
</body>
</html>
